file restructure for sanity
