+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import Routes from "./Routes";
+import Routes from "./containers/Routes";
 import * as serviceWorker from "./serviceWorker";
 
 ReactDOM.render(<Routes />, document.getElementById("root"));
